fix: align direct policy with isolated runtime registry
